Extract FrameAnimator to shared component lib
