+++ desktop/components/frame_animator/easing.js
